backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai
chromadb==0.4.15
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0